requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.7
DateTime==4.3
idna==3.3
lxml==4.6.4
numpy==1.21.4
pandas==1.3.4
pandas-datareader==0.10.0
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
six==1.16.0
urllib3==1.26.7
zope.interface==5.4.0
